third_party/WebKit/LayoutTests/inspector/sources/outline-javascript-4-expected.txt

Verify javascript outline


Running: testAsyncFunction
Text:
    async function foo() { }
Outline:
    foo():0:15

Running: testGeneratorFunction
Text:
    function* foo() { }
Outline:

Running: testMismatchBrackets
Text:
    
    function foo(a, b) {
        if (a > b) {
            return a;
    }
    
    function bar(eee) {
        foo(eee, 2 * eee);
    }
    
Outline:
    foo(a, b):1:9
    bar(eee):6:9

Running: testSyntaxError
Text:
    
    function notAGenerator(a, b) {
        yield 10;
    }
Outline:
    notAGenerator(a, b):1:9
